+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Version 3.3.0 (in progress)
  - [NEW] Added possibility skip server certification validation (`setInsecure()` method)
  - [NEW] Added possibility to query flux on InfuxDB 1.8 using V1 approach
+ - [NEW] `validateConnection()` can be used also for the [forward compatibility](https://docs.influxdata.com/influxdb/latest/tools/api/#influxdb-2-0-api-compatibility-endpoints) connection to InfluxDB 1.8
  - [FIX] Debug compilation error
  - [FIX] SecureBatchWrite compile error